+++ src/components/ProgressButton/ProgressButton.types.ts
@@ -1,0 +1,5 @@
+export interface ProgressButtonProps {
+	onFinish: () => void;
+	title?: string;
+	disabled?: boolean;
+}
