+++ src/components/index.ts
@@ -22,6 +22,7 @@ import PercChange from './PercChange/PercChange';
 import PaperTouchable from './PaperTouchable/PaperTouchable';
 import ApplePayButton from './ApplePayButton/ApplePayButton';
 import ScreenLoadingIndicator from './ScreenLoadingIndicator/ScreenLoadingIndicator';
+import ProgressButton from './ProgressButton/ProgressButton';
 
 export {
 	ActivityIndicator,
@@ -47,5 +48,6 @@ export {
 	ApplePayButton,
 	ScreenLoadingIndicator,
 	Transaction,
-	TransactionIcon
+	TransactionIcon,
+	ProgressButton
 };
