+++ src/screens/DepositScreen/Deposit/Deposit.tsx
@@ -2,12 +2,11 @@ import React, { useEffect } from 'react';
 import { View, TouchableOpacity } from 'react-native';
 import { BasicLayout } from '@layouts';
 import { ParaswapToken } from '@models/token';
-import { Icon, Modal, Text } from '@components';
+import { Icon, Modal, Text, ProgressButton } from '@components';
 import { Card } from 'react-native-paper';
 import { useTheme, useNavigation, useAmplitude } from '@hooks';
 import { debounce } from 'lodash';
 import Warning from '@src/screens/ExchangeScreen/Warning/Warning';
-import ProgressButton from '@src/components/ProgressButton';
 import TransactionWaitModal from '@src/components/TransactionWaitModal/TransactionWaitModal';
 import KeyboardSpacer from 'react-native-keyboard-spacer';
 import TokenCard from '../../ExchangeScreen/TokenCard/TokenCard';
