+++ src/screens/ExchangeResumeScreen/ExchangeResumeScreen.tsx
@@ -15,9 +15,8 @@ import { formatUnits } from 'ethers/lib/utils';
 import { globalWalletState } from '@stores/WalletStore';
 import { BasicLayout } from '@layouts';
 import { TokenType } from '@styles';
-import { Icon, Modal, Text, Token, ActivityIndicator } from '@components';
+import { Icon, Modal, Text, Token, ActivityIndicator, ProgressButton } from '@components';
 import { tokenBalanceFormat } from '@helpers/utilities';
-import ProgressButton from '../../components/ProgressButton';
 import { RootStackParamList } from '../../routes/types.routes';
 import GasOption from '../ExchangeScreen/GasSelector/GasOption/GasOption';
 import { makeStyles } from './ExchangeResume.styles';
